back/requirements.txt
Django>=5.2.3
djangorestframework>=3.16.0
django-cors-headers>=4.3.1
django-filter>=23.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
celery>=5.3.0
redis>=4.5.5
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
django-redis>=5.4.0
djangorestframework-simplejwt>=5.3.1
requests>=2.31.0
aiohttp>=3.9.0
django-phonenumber-field>=7.3.0
phonenumbers>=8.13.30 
websockets>=11.0.3
pandas>=2.0.0
numpy>=1.24.3
ta>=0.10.2
httpx>=0.27.0
joblib>=1.3.2
xgboost>=2.0.3
scikit-learn>=1.5.1

